M3: added the necessary DTOs for the USER database functionality #30
