+++ src/main/java/ch/uzh/ifi/hase/soprafs24/repository/UserRepository.java
@@ -6,7 +6,5 @@ import org.springframework.stereotype.Repository;
 
 @Repository("userRepository")
 public interface UserRepository extends JpaRepository<User, Long> {
-  User findByName(String name);
 
-  User findByUsername(String username);
 }
