+++ src/main/java/ch/uzh/ifi/hase/soprafs24/rest/dto/UserGetDTO.java
@@ -5,9 +5,7 @@ import ch.uzh.ifi.hase.soprafs24.constant.UserStatus;
 public class UserGetDTO {
 
   private Long id;
-  private String name;
-  private String username;
-  private UserStatus status;
+
 
   public Long getId() {
     return id;
@@ -17,27 +15,4 @@ public class UserGetDTO {
     this.id = id;
   }
 
-  public String getName() {
-    return name;
-  }
-
-  public void setName(String name) {
-    this.name = name;
-  }
-
-  public String getUsername() {
-    return username;
-  }
-
-  public void setUsername(String username) {
-    this.username = username;
-  }
-
-  public UserStatus getStatus() {
-    return status;
-  }
-
-  public void setStatus(UserStatus status) {
-    this.status = status;
-  }
 }
