+++ src/main/java/ch/uzh/ifi/hase/soprafs24/rest/dto/UserPostDTO.java
@@ -6,6 +6,10 @@ public class UserPostDTO {
 
   private String username;
 
+  private String password;
+
+  private String email;
+
   public String getName() {
     return name;
   }
@@ -21,4 +25,12 @@ public class UserPostDTO {
   public void setUsername(String username) {
     this.username = username;
   }
+
+  public String getPassword() {return password; }
+
+  public void setPassword(String password) {this.password = password; }
+
+  public String getEmail() {return email;}
+
+  public void setEmail(String email) {this.email = email; }
 }
