+++ src/test/java/ch/uzh/ifi/hase/soprafs24/repository/UserRepositoryIntegrationTest.java
@@ -21,29 +21,5 @@ public class UserRepositoryIntegrationTest {
   @Autowired
   private UserRepository userRepository;
 
-  @Test
-  public void findByName_success() {
-    // given
-    User user = new User();
-    user.setName("Firstname Lastname");
-    user.setUsername("firstname@lastname");
-    user.setStatus(UserStatus.OFFLINE);
-    user.setToken("1");
-    user.setPassword("Password");
-    user.setEmail("user.user@email.com");
-    user.setCreationDate(new Date());
 
-    entityManager.persist(user);
-    entityManager.flush();
-
-    // when
-    User found = userRepository.findByName(user.getName());
-
-    // then
-    assertNotNull(found.getId());
-    assertEquals(found.getName(), user.getName());
-    assertEquals(found.getUsername(), user.getUsername());
-    assertEquals(found.getToken(), user.getToken());
-    assertEquals(found.getStatus(), user.getStatus());
-  }
 }
