+++ src/test/java/ch/uzh/ifi/hase/soprafs24/service/UserServiceIntegrationTest.java
@@ -34,45 +34,5 @@ public class UserServiceIntegrationTest {
     userRepository.deleteAll();
   }
 
-  @Test
-  public void createUser_validInputs_success() {
-    // given
-    assertNull(userRepository.findByUsername("testUsername"));
 
-    User testUser = new User();
-    testUser.setName("testName");
-    testUser.setUsername("testUsername");
-    testUser.setEmail("user.users@gmail.com");
-    testUser.setPassword("password");
-
-    // when
-    User createdUser = userService.createUser(testUser);
-
-    // then
-    assertEquals(testUser.getId(), createdUser.getId());
-    assertEquals(testUser.getName(), createdUser.getName());
-    assertEquals(testUser.getUsername(), createdUser.getUsername());
-    assertNotNull(createdUser.getToken());
-    assertEquals(UserStatus.OFFLINE, createdUser.getStatus());
-  }
-
-  @Test
-  public void createUser_duplicateUsername_throwsException() {
-    assertNull(userRepository.findByUsername("testUsername"));
-
-    User testUser = new User();
-    testUser.setName("testName");
-    testUser.setUsername("testUsername");
-    User createdUser = userService.createUser(testUser);
-
-    // attempt to create second user with same username
-    User testUser2 = new User();
-
-    // change the name but forget about the username
-    testUser2.setName("testName2");
-    testUser2.setUsername("testUsername");
-
-    // check that an error is thrown
-    assertThrows(ResponseStatusException.class, () -> userService.createUser(testUser2));
-  }
 }
